Add "AppName" argument to "notifyService". Allow turn off cflog.
I needed to be able to call "notifyService()" with an application name.
I added a cfargument that is defaulted to what was being used. This
should not change anything for existing installs.

I also added the ability to turn off the cflogging. We have hundreds of
apps. And we will be doing some "info" logging. And we do NOT want our
log folder filling up with hundreds of large CF logs.